My file would not compile in Terraform.  After 2 hours of troubleshooting, still couldn't figure it out.  After downloading the author's "solution" and comparing, still couldn't figure out.  There were some differences in our files, but it didn't look like the differences had anything to do with the errors I was getting.
Decided to scrap my file, which is now "original_main.txt" and run with
"brikis_main.tf".  I had to add in the variable definitions to Brikis' file,
but after that it compiled, ran, and created the cluster.

This update is to push those 2 new files, and changes, to github.